requirements-pqc.txt
cryptography>=3.4


# Optional: python bindings for liboqs (install after building liboqs)
# pip install git+https://github.com/open-quantum-safe/liboqs-python
